Handle JSON parse errors in watch implementation
This change "handles" JSON parse parse errors in the watch implementation by
ignoring them: the JSONStream implementation ignores parsing errors.

The request might result in invalid JSON if, for example, the HTTP connection
is abruptly terminated (and a partial line is piped to the transform stream).

The code this commit replaces would raise an uncaught exception exception on
parse errors. An alternative approach to this commit would invoke the `done`
callback with parse errors, but that could result in multiple invocations of
the `done` callback.